pyproject.toml
[project]
name = "multi_agent_research_system_mars"
version = "0.1.0"
description = "multi_agent_research_system_mars using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.177.0,<1.0.0",
    "flask>=2.3.0",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.95.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "zhipuai>=2.0.0"
]

[project.scripts]
multi_agent_research_system_mars = "multi_agent_research_system_mars.main:run"
run_crew = "multi_agent_research_system_mars.main:run"
train = "multi_agent_research_system_mars.main:train"
replay = "multi_agent_research_system_mars.main:replay"
test = "multi_agent_research_system_mars.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
